# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"Record_string.string_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"infer_typeofCreatePaymentIntentResponseSchema_": {
				"properties": {
					"error": {
						"properties": {
							"declineCode": {
								"type": "string"
							},
							"param": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"validation_error",
									"card_error",
									"api_error",
									"authentication_error"
								]
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"message",
							"code"
						],
						"type": "object"
					},
					"paymentIntent": {
						"properties": {
							"refundedAmount": {
								"type": "number",
								"format": "double"
							},
							"refunded": {
								"type": "boolean"
							},
							"fees": {
								"type": "number",
								"format": "double"
							},
							"amountReceived": {
								"type": "number",
								"format": "double"
							},
							"metadata": {
								"$ref": "#/components/schemas/Record_string.string_"
							},
							"paymentMethodId": {
								"type": "string"
							},
							"customerId": {
								"type": "string"
							},
							"orderId": {
								"type": "string"
							},
							"updatedAt": {
								"type": "string",
								"format": "date-time"
							},
							"createdAt": {
								"type": "string",
								"format": "date-time"
							},
							"status": {
								"type": "string",
								"enum": [
									"processing",
									"requires_payment_method",
									"requires_confirmation",
									"requires_action",
									"succeeded",
									"canceled",
									"requires_capture"
								]
							},
							"currency": {
								"type": "string"
							},
							"amount": {
								"type": "number",
								"format": "double"
							},
							"clientSecret": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updatedAt",
							"createdAt",
							"status",
							"currency",
							"amount",
							"clientSecret",
							"id"
						],
						"type": "object"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object"
			},
			"CreatePaymentIntentResponse": {
				"$ref": "#/components/schemas/infer_typeofCreatePaymentIntentResponseSchema_"
			},
...
			"PaymentError": {
				"$ref": "#/components/schemas/infer_typeofPaymentErrorSchema_"
			},
			"PaymentErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"$ref": "#/components/schemas/PaymentError"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"infer_typeofCreatePaymentIntentRequestSchema_": {
				"properties": {
					"metadata": {
						"$ref": "#/components/schemas/Record_string.string_"
					},
					"description": {
						"type": "string"
					},
					"automaticPaymentMethods": {
						"properties": {
							"allowRedirects": {
								"type": "string",
								"enum": [
									"never",
									"always"
								]
							},
							"enabled": {
								"type": "boolean"
							}
						},
						"required": [
							"enabled"
						],
						"type": "object"
					},
					"paymentMethodId": {
						"type": "string"
					},
					"customerId": {
						"type": "string"
					},
					"orderId": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"amount": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"orderId",
					"currency",
					"amount"
				],
				"type": "object"
			},
			"CreatePaymentIntentRequest": {
				"$ref": "#/components/schemas/infer_typeofCreatePaymentIntentRequestSchema_"
			},
			"infer_typeofConfirmPaymentResponseSchema_": {
				"properties": {
					"nextAction": {
						"properties": {
							"redirectToUrl": {
								"type": "string"
							},
							"type": {
								"type": "string"
							}
						},
						"required": [
							"type"
						],
						"type": "object"
					},
					"requiresAction": {
						"type": "boolean"
					},
					"error": {
						"properties": {
							"declineCode": {
								"type": "string"
							},
							"param": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"validation_error",
									"card_error",
									"api_error",
									"authentication_error"
								]
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"message",
							"code"
						],
						"type": "object"
					},
					"paymentIntent": {
						"properties": {
							"refundedAmount": {
								"type": "number",
								"format": "double"
							},
							"refunded": {
								"type": "boolean"
							},
							"fees": {
								"type": "number",
								"format": "double"
							},
							"amountReceived": {
								"type": "number",
								"format": "double"
							},
							"metadata": {
								"$ref": "#/components/schemas/Record_string.string_"
							},
							"paymentMethodId": {
								"type": "string"
							},
							"customerId": {
								"type": "string"
							},
							"orderId": {
								"type": "string"
							},
							"updatedAt": {
								"type": "string",
								"format": "date-time"
							},
							"createdAt": {
								"type": "string",
								"format": "date-time"
							},
							"status": {
								"type": "string",
								"enum": [
									"processing",
									"requires_payment_method",
									"requires_confirmation",
									"requires_action",
									"succeeded",
									"canceled",
									"requires_capture"
								]
							},
							"currency": {
								"type": "string"
							},
							"amount": {
								"type": "number",
								"format": "double"
							},
							"clientSecret": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updatedAt",
							"createdAt",
							"status",
							"currency",
							"amount",
							"clientSecret",
							"id"
						],
						"type": "object"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object"
			},
			"ConfirmPaymentResponse": {
				"$ref": "#/components/schemas/infer_typeofConfirmPaymentResponseSchema_"
			},
			"Partial_ConfirmPaymentRequest_": {
				"properties": {
					"paymentIntentId": {
						"type": "string"
					},
					"paymentMethodId": {
						"type": "string"
					},
					"returnUrl": {
						"type": "string"
					},
					"useStripeSdk": {
						"type": "boolean"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"infer_typeofRetrievePaymentIntentResponseSchema_": {
				"properties": {
					"error": {
						"properties": {
							"declineCode": {
								"type": "string"
							},
							"param": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"validation_error",
									"card_error",
									"api_error",
									"authentication_error"
								]
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"message",
							"code"
						],
						"type": "object"
					},
					"paymentIntent": {
						"properties": {
							"refundedAmount": {
								"type": "number",
								"format": "double"
							},
							"refunded": {
								"type": "boolean"
							},
							"fees": {
								"type": "number",
								"format": "double"
							},
							"amountReceived": {
								"type": "number",
								"format": "double"
							},
							"metadata": {
								"$ref": "#/components/schemas/Record_string.string_"
							},
							"paymentMethodId": {
								"type": "string"
							},
							"customerId": {
								"type": "string"
							},
							"orderId": {
								"type": "string"
							},
							"updatedAt": {
								"type": "string",
								"format": "date-time"
							},
							"createdAt": {
								"type": "string",
								"format": "date-time"
							},
							"status": {
								"type": "string",
								"enum": [
									"processing",
									"requires_payment_method",
									"requires_confirmation",
									"requires_action",
									"succeeded",
									"canceled",
									"requires_capture"
								]
							},
							"currency": {
								"type": "string"
							},
							"amount": {
								"type": "number",
								"format": "double"
							},
							"clientSecret": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"updatedAt",
							"createdAt",
							"status",
							"currency",
							"amount",
							"clientSecret",
							"id"
						],
						"type": "object"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object"
			},
			"RetrievePaymentIntentResponse": {
				"$ref": "#/components/schemas/infer_typeofRetrievePaymentIntentResponseSchema_"
			},
			"infer_typeofListPaymentMethodsResponseSchema_": {
				"properties": {
					"error": {
						"properties": {
							"declineCode": {
								"type": "string"
							},
							"param": {
								"type": "string"
							},
							"type": {
								"type": "string",
								"enum": [
									"validation_error",
									"card_error",
									"api_error",
									"authentication_error"
								]
							},
							"message": {
								"type": "string"
							},
							"code": {
								"type": "string"
							}
						},
						"required": [
							"type",
							"message",
							"code"
						],
						"type": "object"
					},
					"hasMore": {
						"type": "boolean"
					},
					"paymentMethods": {
						"items": {
							"properties": {
								"accountLast4": {
									"type": "string"
								},
								"bankName": {
									"type": "string"
								},
								"expiryYear": {
									"type": "number",
									"format": "double"
								},
								"expiryMonth": {
									"type": "number",
									"format": "double"
								},
								"brand": {
									"type": "string"
								},
								"last4": {
									"type": "string"
								},
								"isDefault": {
									"type": "boolean"
								},
								"updatedAt": {
									"type": "string",
									"format": "date-time"
								},
								"createdAt": {
									"type": "string",
									"format": "date-time"
								},
								"type": {
									"type": "string",
									"enum": [
										"card",
										"bank_transfer",
										"sepa_debit"
									]
								},
								"id": {
									"type": "string"
								}
							},
							"required": [
								"updatedAt",
								"createdAt",
								"type",
								"id"
							],
							"type": "object"
						},
						"type": "array"
					},
					"success": {
						"type": "boolean"
					}
				},
				"required": [
					"success"
				],
				"type": "object"
			},
			"ListPaymentMethodsResponse": {
				"$ref": "#/components/schemas/infer_typeofListPaymentMethodsResponseSchema_"
			},
...
			"MetalResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"symbol": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetalsListResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"properties": {
							"pagination": {
								"$ref": "#/components/schemas/PaginationResponse"
							},
							"items": {
								"items": {
									"$ref": "#/components/schemas/MetalResponse"
								},
								"type": "array"
							}
						},
						"required": [
							"pagination",
							"items"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetalsErrorResponse": {
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"details": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetalRequest": {
				"properties": {
					"metalName": {
						"type": "string"
					}
				},
				"required": [
					"metalName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "goldsphere-server",
		"version": "1.0.0",
		"description": "The server including the business logic and integration for the goldsphere project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/payments/intent": {
			"post": {
				"operationId": "CreatePaymentIntent",
				"responses": {
					"201": {
						"description": "Payment intent created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/CreatePaymentIntentResponse"
										},
										{
											"$ref": "#/components/schemas/PaymentErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a payment intent",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentIntentRequest"
							}
						}
					}
				}
			}
		},
		"/payments/intent/{id}/confirm": {
			"post": {
				"operationId": "ConfirmPayment",
				"responses": {
					"200": {
						"description": "Payment confirmed",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ConfirmPaymentResponse"
										},
										{
											"$ref": "#/components/schemas/PaymentErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					}
				},
				"description": "Confirm a payment intent",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_ConfirmPaymentRequest_"
							}
						}
					}
				}
			}
		},
		"/payments/intent/{id}": {
			"get": {
				"operationId": "RetrievePaymentIntent",
				"responses": {
					"200": {
						"description": "Payment intent retrieved",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/RetrievePaymentIntentResponse"
										},
										{
											"$ref": "#/components/schemas/PaymentErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					}
				},
				"description": "Retrieve a payment intent",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payments/methods": {
			"get": {
				"operationId": "ListPaymentMethods",
				"responses": {
					"200": {
						"description": "Payment methods retrieved",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ListPaymentMethodsResponse"
										},
										{
											"$ref": "#/components/schemas/PaymentErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid request data",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/PaymentErrorResponse"
								}
							}
						}
					}
				},
				"description": "List payment methods for a customer",
				"tags": [
					"Payments"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "customerId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/metals": {
			"get": {
				"operationId": "ListMetals",
				"responses": {
					"200": {
						"description": "Metals list",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateMetal",
				"responses": {
					"201": {
						"description": "Metal created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetalRequest"
							}
						}
					}
				}
			}
		},
		"/metals/{id}": {
			"put": {
				"operationId": "UpdateMetal",
				"responses": {
					"200": {
						"description": "Metal updated",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetalRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMetal",
				"responses": {
					"204": {
						"description": "Metal deleted"
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetMetalById",
				"responses": {
					"200": {
						"description": "Metal details",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"$ref": "#/components/schemas/MetalResponse"
										},
										"success": {
											"type": "boolean",
											"enum": [
												true
											],
											"nullable": false
										}
									},
									"required": [
										"data",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Metal not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MetalsErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"References"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	],
...
}